feat(@clayui/core): adds the implementation of modifying the Tree by copying the partial structure
The strategy here is to make a partial copy of the structure to deal with the immutable tree but reusing the paths that are not modified, only the paths that are modified undergo changes in the tree that automatically create pointers to the original tree.

The cost is significantly less compared to creating a copy of the tree and then modifying it because we would need to traverse the entire tree to create a copy. We also avoid creating a partial copy before modifying, we create the partial copy in the path iterable.